fix compilation error importing IScope from angular in utils was crashing it
